requirements.txt
# Provider do komunikacji Airflow ze Snowflake
apache-airflow-providers-snowflake

# Biblioteka do uruchamiania dbt jako osobnych tasków w Airflow
astronomer-cosmos

# Snowflake connector
snowflake-connector-python

# Kafka
kafka-python

# Dodatkowe
pydantic
requests
python-dotenv
